Enhance linting, simplify imports, and update dependency packages
This update makes several structural and dependency-based modifications to the code base. ESLint and prettier were configured, streamlining the linting process and enforcing more consistent and cleaner code formatting. Also, import statements in various files were simplified for better readability and maintainability.

Redux was also updated alongside the implementation of babel-eslint-parser, providing access to the latest ECMAScript features and ensuring backward compatibility. Moreover, a list of specified directories and files was added to the ESLint and Prettier ignore files to prevent unnecessary linting and formatting, thus optimizing the application build time.

In addition, several updates to dependencies were carried out to ensure compatibility with React 18 and the latest versions of other libraries.

Modifications were also made to package.json to include a comprehensive set of Babel plugins that enable usage of the latest JavaScript syntax and features. The React version was set to auto-detect, ensuring optimal compatibility while preventing versioning issues.

Lastly, we transitioned to React 18's createRoot API in the root index file. Our application is now fully prepared to take advantage of the upcoming improvements in React 18.