Implement cross-env for build script and adjust dependencies
Transitioned the explicit setting of NODE_ENV in the build script to use cross-env package for enhanced cross-platform compatibility. This change will ensure the correct handling of environmental variables across diverse development environments, eliminating potential errors due to system differences.

Moved "@babel/eslint-parser" from dependencies to devDependencies as it is predominantly used during development and not needed in the production environment. This contributes to a leaner production build.